Separate 'test' command definitions from implementations
